Some bug fixes.

+++ Zbang.Cloudents.Mvc4WebRole/Js/components/flashcard/flashcardCreate.controller.js
@@ -177,6 +177,9 @@ var app;
         FlashcardCreateController.prototype.publish = function () {
             var self = this;
             if (this.form.$invalid) {
+                if (this.form["name"].$invalid) {
+                    this.form["name"].$setTouched();
+                }
                 return;
             }
             else {
@@ -237,7 +240,7 @@ var app;
         };
         FlashcardCreateController.prototype.close = function (ev) {
             var _this = this;
-            if (!this.data.id) {
+            if (!this.data.id && !this.form.$dirty) {
                 this.navigateBackToBox();
                 return;
             }
@@ -248,7 +251,12 @@ var app;
                 .ok(this.resManager.get('quizDelete'))
                 .cancel(this.resManager.get('quizSaveAsDraft'));
             this.$mdDialog.show(confirm).then(function () {
-                _this.flashcardService.delete(_this.data.id).then(_this.navigateBackToBox);
+                if (_this.data.id) {
+                    _this.flashcardService.delete(_this.data.id).then(_this.navigateBackToBox);
+                }
+                else {
+                    _this.navigateBackToBox();
+                }
             }, function () {
                 _this.create().then(_this.navigateBackToBox);
             });
